034_full_deployment/docker-image/requirements.txt
djangorestframework
psycopg2-binary